Fix EOF error on binary STL files
Whether reading as ASCII or binary, the STL file was opened as a text
file. On some systems (like Windows) that means that if the EOF byte was
encountered, the stream would report end of file in the middle. Fix by
making sure the file is opened as binary as necessary.