Add dummy empty array to MPI_Group_excl
We are using MPI_Group_excl to copy a group by providing an empty array
of ranks to exclude. (The documentation states that this will in fact
copy the group.) Some versions of MPI are fine accepting a NULL pointer
for the array of ranks in this case. Other versions are more picky and
want a valid pointer even if pointing to a zero-length array.