Template base image classes using feature structs
Previously, the base implementation classes for images, ImageColorDepth
and ImageColorOnly, would be templated on some base information about
the color and depth format. This was changed to template this base
implementation class on a "feature" structure that contains all this
information and more. Although this puts a bit more trust on the
implementer providing exactly what is needed in the feature struct, it
allows us to easily put a lot more information there. That allows the
base implementation to directly implement methods like getColor. It also
simplifies things like clear so you don't have to jump through hoops
trying to figure out how to encode the color and depth.