Fix issue with blend running past run length
Originally when blending sparse images, we only checked to see if the
top run length was at the end with the assumption that if one runlength
is at the end, they both are. However, if we have received a runlength
from another process, we created a large buffer of runlengths that is
zeroed out, so one of the images might have a bunch of runlengths of
size zero at the end. Thus, one image might be at the end of its
runlengths while the other has a bunch of 0 lengths at its end.

Fix the problem by checking both run lengths for the end. Assume that if
one is at the end, that the other has only 0 size runlengths and we can
quit.