Support blending of images with different regions
These changes allow you to take images that only partially overlap and
generate a composite image containing a blend of the overlapping parts
and a copy of the non-overlapping parts. This adds quite a bit of
complication to this part of the code, but I think it will be necessary
to implement 2-3 swap.