Fix blending of clear sparse images of different regions
Previously whenever you blended two sparse images and one had no active
pixels, blend just returned a deep copy of the non-clear image. However,
this is an incorrect result of the two regions are not the same (and the
resulting image should be the entire region). Thus, add a check that the
regions are the same before using this fast path.