Prevent getting stuck when traversing sparse run lengths
The run length iterator had a bug where it was possible to get "stuck"
in a situation where both the background and foreground pixels were zero
but it was not at the end (i.e. it needed to go to the next run length).
This usually causes code to get stuck because it doesn't know what type
the next pixel is to advance. Fix this issue and also add some more
checks.