Allow any image to be saved to a PPM
Previously, you could only save an ImageFull to a PPM, and that would
crash if the image was actually a subregion. Now, images with a narrow
region are expanded to be written out as a full image and sparse images
are uncompressed. Now it is a lot easer to write out images.