Fix issue with incomplete incoming images
After waiting for the "last" incoming message for an image in direct
send overlap, I neglected to wait for the remaining images. Apparently
it is possible for messages to complete out of order.

Fixed by calling MPI_Waitall on the remaining images (which should be
done anyway to free up the objects).