Make distribution of blocks more square
Previously when broadcasting blocks, the blocks were placed in an n^3
grid where n is the smallest integer in which n^3 is at least as large
as the number of blocks/processes. The problem is that when the number
of processes falls just past a perfect cube, the blocks got distributed
very unevenly. So change this so that the dimensions of this cube can
vary by 1. This places the blocks in a tigher rectangle in many
instances.