Fix C++11 warning about string literals
In C++11 you are not allowed to assign a string literal to a char*
variable. Fix this by copying the string literal. (Note, this will leave
a small buffer that is not free by the termination of the program. Not a
big deal, but a memory analyser might pick that up.)