Fix performance issue with building 2-3 swap composite tree
There was a mistake where a temporary buffer that should have been sized
to the number of processes in a group was sized to the number of pixels
in the image. Although that made the size of the buffer almost certainly
large enough, allocating (and deleting) a buffer that large over and
over again was a significant performance hit.

This now sizes these buffers correctly.