Save images from failed composite in unique file
Previously, whenever a composite failed, it would write a couple of
images to some prescribed files. The problem was that if you had a batch
job with a bunch of compositing, the error images would write over each
other. This fixes the problem by tagging the file with the current time,
which makes it very likely to be unique.