Load miniGraphicsMacros.cmake in top directory to speed config
Previously, running cmake on miniGraphics was quite slow. I narrowed
down the problem to the fact that find(MPI) was called in
miniGraphicsMacros.cmake, and finding MPI can be pretty slow. If you
find MPI once that is not a huge deal, but it was re-finding MPI for
every miniapp subdirectory.

However, if MPI is already found, then find(MPI) returns immediately
with the previously found configuration. Thus, a simple solution is to
include miniGraphicsMacros.cmake at the top level. Then all the
subdirectories will inherit the configuration it finds and can skip the
slow package finding in the subdirectories.