Use first/second instead of top/bottom for image splits
Previously, binary swap and similar algorithms used the terms top and
bottom when referring to the halves of a split image. These were a bit
of a misnomer because "top" was actually in the visible bottom of the
image (although at the beginning of the memory buffer). To avoid
confusion, the names are changed to first/second.